Organize the remaining files
