+++ compiler/core/src/main.bs.js
@@ -9,7 +9,7 @@ var $$Array                              = require("bs-platform/lib/js/array.js"
 var Process                              = require("process");
 var FsExtra                              = require("fs-extra");
 var Caml_array                           = require("bs-platform/lib/js/caml_array.js");
-var Color$LonaCompilerCore               = require("./color.bs.js");
+var Color$LonaCompilerCore               = require("./core/color.bs.js");
 var SwiftColor$LonaCompilerCore          = require("./swift/swiftColor.bs.js");
 var SwiftRender$LonaCompilerCore         = require("./swift/swiftRender.bs.js");
 var SwiftComponent$LonaCompilerCore      = require("./swift/swiftComponent.bs.js");
