+++ compiler/core/src/utils/css.bs.js
@@ -45,7 +45,7 @@ function parseColor(value) {
       throw [
             Caml_builtin_exceptions.match_failure,
             [
-              "/Users/devin_abbott/Projects/ComponentStudio/ComponentStudio/compiler/core/src/css.re",
+              "/Users/devin_abbott/Projects/ComponentStudio/ComponentStudio/compiler/core/src/utils/css.re",
               15,
               8
             ]
